Fix AQI_missions.txt
In the current version, forming Gallia as Aquitaine prevents most Aquitaine mission to be completable, which is probably not intended (at leasts makes the game weird as you don't want to form Gallia as soon as you can). This proposed modification (which amounts to change most AQI tags to ROOT tags in triggers and effect) should fix that as missions would remain doable when switching from AQI to GLI, while not changing anything as long as playing as AQI.
Note: I noticed that through a playthrought with Aquitaine. You might want to check if there are not other countries where there are similar mistakes.